move auth-related files into auth subfolder
