+++ apps/web/src/lib/server/auth/password.ts
@@ -1,7 +1,7 @@
-import { Bindings } from './bindings';
+import { Bindings } from '$lib/server/bindings';
 import { sha1 } from '@oslojs/crypto/sha1';
 import { encodeHexLowerCase } from '@oslojs/encoding';
-import { ExpiringTokenBucketProxy } from './rate-limit/ExpiringTokenBucketProxy';
+import { ExpiringTokenBucketProxy } from '$lib/server/rate-limit/ExpiringTokenBucketProxy';
 
 export const getPasswordUpdateBucket = async (userId: string) => {
 	return ExpiringTokenBucketProxy.initialize({
