+++ apps/web/src/routes/(auth)/2fa/+page.server.ts
@@ -1,8 +1,8 @@
-import { getTotpBucket } from '$lib/server/2fa';
+import { getTotpBucket } from '$lib/server/auth/2fa';
 import { fail, redirect } from '@sveltejs/kit';
-import { getUserTOTPKey } from '$lib/server/user';
+import { getUserTOTPKey } from '$lib/server/auth/user';
 import { verifyTOTP } from '@oslojs/otp';
-import { setSessionAs2FAVerified } from '$lib/server/session';
+import { setSessionAs2FAVerified } from '$lib/server/auth/session';
 
 import type { Actions, RequestEvent } from './$types';
