+++ apps/web/src/routes/(auth)/2fa/reset/+page.server.ts
@@ -1,4 +1,4 @@
-import { getRecoveryCodeBucket, resetUser2FAWithRecoveryCode } from '$lib/server/2fa';
+import { getRecoveryCodeBucket, resetUser2FAWithRecoveryCode } from '$lib/server/auth/2fa';
 import { fail, redirect } from '@sveltejs/kit';
 
 import type { Actions, RequestEvent } from './$types';
