+++ apps/web/src/routes/(auth)/oauth/github/+server.ts
@@ -1,5 +1,5 @@
 import { generateState } from 'arctic';
-import { github } from '$lib/server/oauth';
+import { github } from '$lib/server/auth/oauth';
 
 import type { RequestEvent } from '@sveltejs/kit';
