+++ apps/web/src/routes/(auth)/register/+page.server.ts
@@ -5,15 +5,15 @@ import {
 	generateSessionToken,
 	setSessionTokenCookie,
 	type SessionFlags
-} from '$lib/server/session';
-import { createUser } from '$lib/server/user';
-import { verifyPasswordStrength } from '$lib/server/password';
-import { checkEmailAvailability, verifyEmailInput } from '$lib/server/email';
+} from '$lib/server/auth/session';
+import { createUser } from '$lib/server/auth/user';
+import { verifyPasswordStrength } from '$lib/server/auth/password';
+import { checkEmailAvailability, verifyEmailInput } from '$lib/server/auth/email';
 import {
 	createEmailVerificationRequest,
 	sendVerificationEmail,
 	setEmailVerificationRequestCookie
-} from '$lib/server/email-verification';
+} from '$lib/server/auth/email-verification';
 import { RefillingTokenBucketProxy } from '$lib/server/rate-limit/RefillingTokenBucketProxy';
 
 export const load: PageServerLoad = async (event) => {
