+++ apps/web/src/routes/(auth)/reset-password/2fa/+page.server.ts
@@ -1,12 +1,12 @@
 import { verifyTOTP } from '@oslojs/otp';
-import { getUserTOTPKey } from '$lib/server/user';
+import { getUserTOTPKey } from '$lib/server/auth/user';
 import {
 	validatePasswordResetSessionRequest,
 	setPasswordResetSessionAs2FAVerified
-} from '$lib/server/password-reset';
-import { getTotpBucket, getRecoveryCodeBucket } from '$lib/server/2fa';
+} from '$lib/server/auth/password-reset';
+import { getTotpBucket, getRecoveryCodeBucket } from '$lib/server/auth/2fa';
 import { fail, redirect } from '@sveltejs/kit';
-import { resetUser2FAWithRecoveryCode } from '$lib/server/2fa';
+import { resetUser2FAWithRecoveryCode } from '$lib/server/auth/2fa';
 
 import type { Actions, RequestEvent } from './$types';
