+++ apps/web/src/routes/(auth)/reset-password/verify-email/+page.server.ts
@@ -2,8 +2,8 @@ import {
 	validatePasswordResetSessionRequest,
 	setPasswordResetSessionAsEmailVerified,
 	getVerifyEmailBucket
-} from '$lib/server/password-reset';
-import { setUserAsEmailVerifiedIfEmailMatches } from '$lib/server/user';
+} from '$lib/server/auth/password-reset';
+import { setUserAsEmailVerifiedIfEmailMatches } from '$lib/server/auth/user';
 import { fail, redirect } from '@sveltejs/kit';
 
 import type { Actions, RequestEvent } from './$types';
