+++ apps/web/src/routes/(protected)/+page.server.ts
@@ -1,6 +1,6 @@
 import { fail, redirect } from '@sveltejs/kit';
 import type { Actions, PageServerLoad } from './$types';
-import { invalidateSession, deleteSessionTokenCookie } from '$lib/server/session';
+import { invalidateSession, deleteSessionTokenCookie } from '$lib/server/auth/session';
 
 export const load: PageServerLoad = async (event) => {
 	// if (event.locals.session === null || event.locals.user === null) {
