+++ apps/web/src/worker-configuration.d.ts
@@ -8,6 +8,7 @@ interface Env {
 		import('@ficus/rate-limiter/src/index').ExpiringTokenBucket
 	>;
 	THROTTLER: DurableObjectNamespace<import('@ficus/rate-limiter/src/index').Throttler>;
+	DOCUMENTS: R2Bucket;
 	DB: D1Database;
 	ARGON2: Fetcher;
 }
